refactored utility functions to utils.rs
